feat(tests): consolidate mock imports for improved test organization
